GlmEstimates-class {snpStats}	R Documentation

Class "GlmEstimates"

Description

A simple class to hold output from snp.lhs.estimates and snp.rhs.estimates. Its main purpose is to provide a show method

Objects from the Class

Objects from this class are simple lists. Each element of the list is a list giving the results of a generalized linear model fit, with elements:

Y.var

Name of the Y variable

beta

The vector or parameter estimates (with their names)

Var.beta

The upper triangle of the variance-covariance matrix of estimates, stored as a simple vector

N

The number of "units" used in the model fit

Extends

Class "list", from data part. Class "vector", by class "list", distance 2.

Methods

[

signature(x = "GlmEstimates", i = "ANY", j = "missing", drop = "missing"): Subset

coerce

signature(from = "GlmEstimates", to = "GlmTests"): Calculate Wald tests

show

signature(object = "GlmEstimates"): Display

Note

Wald tests calculated by coercing an object of class GlmEstimates to class GlmTests are asymptotically equivalent to the "score" tests calculated by snp.lhs.tests and snp.rhs.tests, and both types of test are asymptotically correct. But the asymptotic properties of Wald tests are not as good as those of score tests and, in circumstances such as low sample size or low minor allele frequency, Wald and score tests may differ substantially. In general score tests are to be preferred, but Wald tests are provided in snpStats because they are widely used.

Author(s)

David Clayton dc208@cam.ac.uk

See Also

snp.lhs.estimates, snp.rhs.estimates

Examples

showClass("GlmEstimates")

[Package snpStats version 1.44.0 Index]