GlmTests-class {snpStats}	R Documentation

Classes "GlmTests" and "GlmTestsScore"

Description

Classes of objects created by snp.lhs.tests and snp.rhs.tests. The class "GlmTestsScore" extends the class "GlmTests" and is invoked by setting the argument score=TRUE when calling testing functions in order to save the scores and their variances (and covariances)

Objects from the Class

Objects of class "GlmTests" have four slots:

snp.names

When only single SNPs are tested, a character vector of SNP names. Otherwise a list of such vectors (one for each test)

var.names

A character vector containing names of variables tested against SNPs

chisq

A numerical vector of chi-squared test values

df

An integer vector of degrees of freedom for the tests

N

A integer vector of the number of samples contributing to each test

The "GlmTestsScore" class extends this, adding a slot score containing a list with elements which are themselves lists with two elements:

U

The vector (or matrix) of efficient scores

V

The upper triangle of the variance-covariance matrix of U, stored as a vector

Methods

[]

signature(x = "GlmTests", i = "ANY", j = "missing", drop = "missing"): Subsetting operator

coerce

signature(from = "GlmTests", to = "data.frame"): Simplify object

chi.squared

signature(x = "GlmTests", df = "missing"): Extract chi-squared test values

deg.freedom

signature(x = "GlmTests"): Extract degrees of freedom for tests

names

signature(x="GlmTests"): Extract (or generate) a name for each test

p.value

signature(x = "GlmTests", df = "missing"): Extract p-values

sample.size

signature(object = "GlmTests"): Extract sample sizes for tests

show

signature(object = "GlmTests"): Show method

summary

signature(object = "GlmTests"): Summary method

[]

signature(x = "GlmTestsScore", i = "ANY", j = "missing", drop = "missing"): Subsetting operator

effect.sign

signature(x = "GlmTestsScore", simplify = "logical"): Extract signs of associations. If simpify is TRUE then a simple vector is returned if all tests are on 1df

pool2

signature(x = "GlmTestsScore", y = "GlmTestsScore", score = "logical"): Combine results from two sets of tests

switch.alleles

signature(x = "GlmTestsScore", snps = "character"): Emulate, in the score vector and its (co)variances, the effect of switching of the alleles of specified SNPs

Note

Most of the methods for this class are shared with the SingleSnpTests and SingleSnpTestsScore classes

Author(s)

David Clayton dc208@cam.ac.uk

See Also

snp.lhs.tests,snp.rhs.tests, SingleSnpTests, SingleSnpTestsScore

Examples

showClass("GlmTests")

[Package snpStats version 1.44.0 Index]