ImputationRules-class {snpStats}	R Documentation

Class "ImputationRules"

Description

A class defining a list "rules" for imputation of SNPs. Rules are estimated population haplotype probabilities for a target SNP and one or more predictor SNPs

Objects from the Class

Objects are lists of rules. Rules are named list elements each describing imputation of a SNP by a linear regression equation. Each element is itself a list with the following elements:

maf

The minor allele frequency of the imputed SNP

r.squared

The squared Pearson correlation coefficient between observed and predicted SNP duration derivation of the rule.

snps

The names of the SNPs to be included in the regression.

hap.probs

A numerical array containing estimated probabilities for haplotypes of the SNP to be imputed and all the predictor SNPs

If any target SNP is monomorphic, the corresponding rule is returned as NULL. An object of class ImputationRules has an attribute, Max.predictors, which gives the maximum number of predictors used for any imputation.

Methods

show

signature(object = "ImputationRules"): prints an abreviated listing of the rules

summary

signature(object = "ImputationRules"): returns a table which shows the distribution of r-squared values achieved against the number of snps used for imputation

plot

signature(x="ImputationRules", y="missing"): plots the distribution of r-squared values as a stacked bar chart

[]

signature(x = "ImputationRules", i = "ANY"): subset operations

Author(s)

David Clayton dc208@cam.ac.uk

See Also

snp.imputation, impute.snps, single.snp.tests

Examples

showClass("ImputationRules")

[Package snpStats version 1.44.0 Index]