convert.snpMatrix {snpStats}	R Documentation

Convert snpMatrix objects to snpStats objects

Description

These functions convert snpMatrix objects to snpStats objects. convert.snpMatrix converts a single object, while convert.snpMatrix.dir converts all stored elements in a specified directory. They really only change the class names since most of the classes in snpStats are backwards-compatible with snpMatrix. The exception is the ImputationRules class; imputation.rules objects will need to be regenerated.

Usage

convert.snpMatrix(object)

convert.snpMatrix.dir(dir = ".", ext = "RData")

Arguments

object
Object to be converted

dir
A directory containing saved snpMatrix objects

ext
The file extension for files containing such objects

Value

convert.snpMatrix returns the converted object. convert.snpMatrix.dir rewrites the files in place.

Author(s)

David Clayton dc208@cam.ac.uk

[Package snpStats version 1.44.0 Index]