families {snpStats}	R Documentation

Test data for family association tests

Description

These data started life as real data derived from an affected sibling pair study of type 1 diabetes. However, original subject and SNP identidiers have been replaced by randomly chosen ones.

Usage

data(families)

Format

There are two objects in the loaded data file:

genotypes: An object of class "snp.matrix" containing the SNP genotype data for both parents and affected offspring

pedData: A data frame containing the standard six fields for a LINKAGE pedfile. The are named familyid, member, father, mother sex, and affected

The two objects are linked by common row names.

Details

Coding in the pedData frame is as in the LINKAGE package, except that missing data are coded NA rather than zero

Examples

data(families)
summary(genotypes)
summary(pedData)

[Package snpStats version 1.44.0 Index]