filter.rules {snpStats}	R Documentation

Filter a set of imputation rules

Description

Determine which imputation rules are broken by removal of some SNPs from a study. This function is needed because, when if it emerges that genotyping of some SNPs is not reliable, necessitating their removal from study, we would also wish to remove any SNPs imputed on the basis of these unreliable SNPs.

Usage

filter.rules(rules, snps.excluded, exclusions = TRUE)

Arguments

rules
An object of class "ImputationRules" containing a set of imputation rules

snps.excluded
The names of the SNPs whose removal is to be investigated

exclusions
If TRUE, the names of the imputed SNPs which would be lost by removal of the SNPs listed in snps.excluded. If FALSE, the names of the imputed SNPs which would not be lost are returned

Value

A character vector containing the names of imputed SNPs to be removed

Author(s)

David Clayton dc208@cam.ac.uk

See Also

ImputationRules-class, snp.imputation

Examples

# No example yet

[Package snpStats version 1.44.0 Index]