ibsDist {snpStats}	R Documentation

Distance matrix based on identity by state (IBS)

Description

Expresses a matrix of IBS counts (see ibsCount) as a distance matrix. The distance between two samples is returned as the proportion of allele comparisons which are not IBS.

Usage

ibsDist(counts)

Arguments

counts
A matrix of IBS counts as produced by the function ibsCount

Value

An object of class "dist" (see dist)

Author(s)

David Clayton dc208@cam.ac.uk

See Also

ibsCount, dist

Examples

data(testdata)
ibs <- ibsCount(Xchromosome)
distance <- ibsDist(ibs)

[Package snpStats version 1.44.0 Index]