imputation.maf {snpStats}	R Documentation

Extract statistics from imputation rules

Description

These functions extract key characteristics of regression-based imputation rules stored as an object of class "ImputationRules". imputation.maf extracts the minor allele frequencies of the imputed SNPs and imputation.r2 extracts the prediction R^2.

Usage

can.impute(rules)
imputation.maf(rules)
imputation.r2(rules)
imputation.nsnp(rules)

Arguments

rules
An object of class "ImputationRules"

Details

can.impute returns a logical vector identifying which rules allow a valid imputation. imputation.maf and imputation.r2 extract the minor allele frequencies of the imputed SNPs and the R^2 for prediction achieved when building each rule. imputation.nsnp returns the numbers of SNPs used in each imputation

Value

Either a logical vector, or a numeric vector containing the extracted values

Author(s)

David Clayton dc208@cam.ac.uk

See Also

ImputationRules-class, snp.imputation

Examples

# These functions are currently defined as
function (rules) sapply(rules, function(x) x$maf)
function (rules) sapply(rules, function(x) x$r2)

[Package snpStats version 1.44.0 Index]