ld.example {snpStats}	R Documentation

Datasets to illustrate calculation of linkage disequilibrium statistics

Description

This R data file contains data from the International HapMap project, concerning 603 SNPs spanning a one megabase region on chromosome 22, in a sample of Europeans and a sample of Africans

Format

There are three objects in the file:

ceph.1m: A snpMatrix object containing the European genotype data

yri.1m: A snpMatrix object containing the African genotype data

support.ld: A dataframe containing details (chromosome position etc.

of the 603 SNPs

Source

http://hapmap.ncbi.nlm.nih.gov

References

The International HapMap Consortium. The International HapMap Project. Nature 426:789-796 (2003)

[Package snpStats version 1.44.0 Index]