mvtests {snpStats}	R Documentation

Multivariate SNP tests

Description

This function calculates multivariate score tests between a multivariate (or multinomial) phenotype and sets of SNPs

Usage

mvtests(phenotype, sets, stratum, data = sys.parent(), snp.data, rules = NULL, complete = TRUE, uncertain = FALSE, score = FALSE)

Arguments

phenotype
Either a factor (for a multinomial phenotype) or a matrix (for a multivariate phenotype)

sets
A list of sets of SNPs to be tested against the phenotype

stratum
(Optional) a stratifying variable

data
A data frame in which phenotype and stratum reside. If absent these are assumed to be in the parent frame and correctly aligned with the rows of snp.data

snp.data
An object of class SnpMatrix containing the SNP data

rules
(Optional) A set of imputation rules. The function then carries out tess on imputed SNPs

complete
If TRUE each test will use only subjects who have complete data for the phenotype and all SNPs in the set to be tested. If FALSE, then complete data for the phenotype is required, but tests are based upon complete pairs of SNPs

uncertain
If TRUE, uncertain genotype calls will be used in the tests (scored by their posterior expectations). Otherwise such calls are treated as missing

score
If TRUE, the score vectors and their variance-covariance matrices are saved in the output object for further processing

Details

Currently complete=FALSE is not implemented

Value

An object of class snp.tests.glm or GlmTests.score depending on whether score is set to FALSE or TRUE in the call

Note

This is an experimental version

Author(s)

David Clayton dc208@cam.ac.uk

Examples

## No example yet

[Package snpStats version 1.44.0 Index]