pool {snpStats}	R Documentation

Pool test results from several studies or sub-studies

Description

Given the same set of "score" tests carried out in several studies or in several different sub-samples within a study, this function pools the evidence by summation of the score statistics and score variances. It combines tests produced by single.snp.tests or by snp.lhs.tests and snp.rhs.tests.

Usage

pool(..., score = FALSE)

Arguments

...
Objects holding the (extended) test results. These must be of class SingleSnpTests.score or snp.tests.glm

score
Is extended score information to be returned in the output object? Relevant only for SingleSnpTestsScore objects

Details

This function works by recursive calls to the generic function pool2 which pools the results of two studies.

Value

An object of same class as the input objects (optionally without the .score) extension. Tests are produced for the union of SNPs tested in all the input objects.

Author(s)

David Clayton dc208@cam.ac.uk

See Also

pool2, SingleSnpTestsScore-class, GlmTests-class, single.snp.tests, snp.lhs.tests, snp.rhs.tests

Examples

# An artificial example which simply doubles the size of a study
data(testdata)
sst <- single.snp.tests(snp.data=Autosomes, cc, data=subject.data,
         score=TRUE)
sst2 <- pool(sst, sst)
summary(sst2)

[Package snpStats version 1.44.0 Index]