pool2 {snpStats}	R Documentation

Pool results of tests from two independent datasets

Description

Generic function to pool results of tests from two independent datasets. It is not designed to be called directly, but is called recursively by pool

Usage

pool2(x, y, score)

Arguments

x, y
Objects holding the (extended) test results. These must be of class SingleSnpTests.score or snp.tests.glm

score
Is extended score information to be returned in the output object?

Value

An object of same class as the input objects (optionally without the .score) extension. Tests are produced for the union of SNPs tested in all the input objects.

Author(s)

David Clayton dc208@cam.ac.uk

See Also

pool, SingleSnpTestsScore-class, GlmTests-class, single.snp.tests, snp.lhs.tests, snp.rhs.tests

[Package snpStats version 1.44.0 Index]