read.mach {snpStats}	R Documentation

Read genotypes imputed by the MACH program

Description

This routine reads imputed genotypes generated by the MACH program. With the --mle and --mldetails options in force this program generates a .mlprob output file which contains probabilities of assignments. These are stored as uncertain genotype calls in a SnpMatrix object

Usage

read.mach(file, colnames = NULL, nrow = NULL)

Arguments

file
The name of the .mlprob file. This may be gzipped

colnames
The column names. If absent, names are generated as SNP1, SNP2, etc.

nrow
If known the number of rows of data on the file. If not supplied, it is determined by a preliminary pass through the data

Details

No routine is explicitly available for data on chromosome X. Such data should first be read as a SnpMatrix and then coerced to an XSnpMatrix object

Value

An object of class SnpMatrix

Author(s)

David Clayton dc208@cam.ac.uk

See Also

SnpMatrix-class

Examples

##---- No example available yet

[Package snpStats version 1.44.0 Index]