testdata {snpStats}	R Documentation

Test data for the snpStats package

Description

This dataset comprises several data frames from a fictional (and unrealistically small) study. The dataset started off as real data from a screen of non-synonymous SNPs for association with type 1 diabetes, but the original identifiers have been removed and a random case/control status has been generated.

Usage

data(testdata)

Format

There are five data objects in the dataset:

Autosomes: An object of class "SnpMatrix" containing genotype calls for 400 subjects at 9445 autosomal SNPs

Xchromosome: An object of class "XSnpMatrix" containing genotype calls for 400 subjects at 155 SNPs on the X chromosome

Asnps: A dataframe containing information about the autosomal SNPs. Here it contains only one variable, chromosome, indicating the chromosomes on which the SNPs are located

Xsnps: A dataframe containing information about the X chromosome SNPs. Here it is empty and is only included for completeness

subject.data: A dataframe containing information about the subjects from whom each row of SNP data was obtained. Here it contains:

cc: Case-control status

sex: Sex

region: Geographical region of residence

Source

The data were obtained from the diabetes and inflammation laboratory (see http://www-gene.cimr.cam.ac.uk)

Examples

data(testdata)
Autosomes
Xchromosome
summary(Asnps)
summary(Xsnps)
summary(subject.data)
summary(summary(Autosomes))
summary(summary(Xchromosome))

[Package snpStats version 1.44.0 Index]