write.SnpMatrix {snpStats}	R Documentation

Write a SnpMatrix object as a text file

Description

This function is closely modelled on write.table. It writes an object of class SnpMatrix as a text file with one line for each row of the matrix. Genotpyes are written in numerical form, i.e. as 0, 1 or 2 (where 1 denotes heterozygous) or, optionally, as a pair of alleles (each coded 1 or 2).

Usage

write.SnpMatrix(x, file, as.alleles= FALSE, append = FALSE, quote = TRUE, sep = " ", eol = "\n", na = "NA", row.names = TRUE, col.names = TRUE)

Arguments

x
The object to be written

file
The name of the output file

as.alleles
If TRUE, write each genotype as two alleles

append
If TRUE, the output is appended to the designated file. Otherwise a new file is opened

quote
If TRUE, row and column names will be enclosed in quotes

sep
The string separating entries within a line

eol
The string terminating each line

na
The string written for missing genotypes

row.names
If TRUE, each row will commence with the row name

col.names
If TRUE, the first line will contain all the column names

Value

A numeric vector giving the dimensions of the matrix written

Author(s)

David Clayton dc208@cam.ac.uk

See Also

write.table, SnpMatrix-class, XSnpMatrix-class

[Package snpStats version 1.44.0 Index]